+++ magda-storage-api/src/createApiRouter.ts
@@ -30,6 +30,57 @@ export default function createApiRouter(options: ApiRouterOptions) {
     });
 
     /**
+     * @apiGroup Storage
+     *
+     * @api {PUT} /v0/{bucketid} Request to create a new bucket
+     *
+     * @apiDescription Creates a new bucket with a specified name. Restricted to admins only.
+     *
+     * @apiParam (Request body) {string} bucketid The name of the bucket to be created
+     *
+     * @apiSuccessExample {json} 201
+     *    {
+     *        "message":"Bucket my-bucket created successfully in unspecified-region 🎉"
+     *    }
+     *
+     * @apiSuccessExample {json} 201
+     *    {
+     *        "message": "Bucket my-bucket already exists 👍"
+     *    }
+     * @apiErrorExample {json} 500
+     *    {
+     *        "message": "Bucket creation failed. This has been logged and we are looking into this."
+     *    }
+     */
+    router.put(
+        "/:bucketid",
+        mustBeAdmin(options.authApiUrl, options.jwtSecret),
+        async function(req, res) {
+            const bucketId = req.params.bucketid;
+            if (!bucketId) {
+                return res
+                    .status(400)
+                    .send("Please specify a bucket name in the request URL.");
+            }
+
+            const encodedBucketname = encodeURIComponent(bucketId);
+            try {
+                const createBucketRes = await options.objectStoreClient.createBucket(
+                    encodedBucketname
+                );
+                return res.status(201).send({
+                    message: createBucketRes.message
+                });
+            } catch (err) {
+                return res.status(500).send({
+                    message:
+                        "Bucket creation failed. This has been logged and we are looking into this."
+                });
+            }
+        }
+    );
+
+    /**
      * @apiDefine Storage Storage API
      */
 
@@ -98,7 +149,7 @@ export default function createApiRouter(options: ApiRouterOptions) {
      *
      * @api {put} /v0/{bucket}/{fieldid} Request to upload an object to {bucket} with name {fieldid}
      *
-     * @apiDescription Uploads an object
+     * @apiDescription Uploads an object. Restricted to admins only.
      *
      * @apiParam (Request body) {string} bucket The name of the bucket to which to upload to
      * @apiParam (Request body) {string} fieldid The name of the object being uploaded
@@ -106,7 +157,7 @@ export default function createApiRouter(options: ApiRouterOptions) {
      * @apiSuccessExample {json} 200
      *    {
      *        "message":"File uploaded successfully",
-     *        "etag":"edd88378a7900bf663a5fa386396b585-1"
+     *        "etag":"edd88378a7900bf663a5fa386386b585-1"
      *    }
      *
      * @apiErrorExample {json} 500
@@ -123,9 +174,16 @@ export default function createApiRouter(options: ApiRouterOptions) {
             const encodedRootPath = encodeURIComponent(fileId);
             const encodeBucketname = encodeURIComponent(bucket);
             const content = req.body;
+            const contentType = req.headers["content-type"];
+            const contentLength = req.headers["content-length"];
+
+            if (!contentLength) {
+                return res.status(400).send("No Content.");
+            }
+
             const metaData = {
-                "Content-Type": req.headers["content-type"],
-                "Content-Length": req.headers["content-length"]
+                "Content-Type": contentType,
+                "Content-Length": contentLength
             };
             return options.objectStoreClient
                 .putFile(encodeBucketname, encodedRootPath, content, metaData)